moves tests to msgraphcore_test directory and package
