+++ msgraphgocore_test/page_iterator_test.go
@@ -1,4 +1,4 @@
-package msgraphgocore
+package msgraphgocore_test
 
 import (
 	"fmt"
@@ -6,9 +6,11 @@ import (
 	httptest "net/http/httptest"
 	testing "testing"
 
+	msgraphgocore "github.com/microsoftgraph/msgraph-sdk-go-core"
+	"github.com/microsoftgraph/msgraph-sdk-go-core/mocks"
+
 	"github.com/microsoft/kiota/abstractions/go/authentication"
 	"github.com/microsoft/kiota/abstractions/go/serialization"
-	"github.com/microsoftgraph/msgraph-sdk-go-core/msgraphgocore_test"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -21,13 +23,13 @@ type UserPage struct {
 	NextLink *string
 }
 
-var reqAdapter, _ = NewGraphRequestAdapterBase(&authentication.AnonymousAuthenticationProvider{}, GraphClientOptions{
+var reqAdapter, _ = msgraphgocore.NewGraphRequestAdapterBase(&authentication.AnonymousAuthenticationProvider{}, msgraphgocore.GraphClientOptions{
 	GraphServiceVersion:        "",
 	GraphServiceLibraryVersion: "",
 })
 
 func ParsableCons() serialization.Parsable {
-	return msgraphgocore_test.NewUsersResponse()
+	return mocks.NewUsersResponse()
 }
 
 func TestIterateStopsWhenCallbackReturnsFalse(t *testing.T) {
@@ -48,10 +50,10 @@ func TestIterateStopsWhenCallbackReturnsFalse(t *testing.T) {
 
 	}))
 	defer testServer.Close()
-	pageIterator := NewPageIterator(graphResponse, *reqAdapter, ParsableCons, nil)
+	pageIterator := msgraphgocore.NewPageIterator(graphResponse, *reqAdapter, ParsableCons, nil)
 
 	pageIterator.Iterate(func(pageItem interface{}) bool {
-		item := pageItem.(msgraphgocore_test.User)
+		item := pageItem.(mocks.User)
 
 		res = append(res, *item.GetDisplayName())
 		return !(*item.GetId() == "2")
@@ -81,11 +83,11 @@ func TestIterateEnumeratesAllPages(t *testing.T) {
 	mockPath := testServer.URL + "/next-page"
 	graphResponse.SetNextLink(&mockPath)
 
-	pageIterator := NewPageIterator(graphResponse, *reqAdapter, ParsableCons, nil)
+	pageIterator := msgraphgocore.NewPageIterator(graphResponse, *reqAdapter, ParsableCons, nil)
 	res := make([]string, 0)
 
 	pageIterator.Iterate(func(pageItem interface{}) bool {
-		item := pageItem.(msgraphgocore_test.User)
+		item := pageItem.(mocks.User)
 		res = append(res, *item.GetId())
 		return true
 	})
@@ -118,9 +120,9 @@ func TestIterateCanBePausedAndResumed(t *testing.T) {
 	mockPath := testServer.URL + "/next-page"
 	response.SetNextLink(&mockPath)
 
-	pageIterator := NewPageIterator(response, *reqAdapter, ParsableCons, nil)
+	pageIterator := msgraphgocore.NewPageIterator(response, *reqAdapter, ParsableCons, nil)
 	pageIterator.Iterate(func(pageItem interface{}) bool {
-		item := pageItem.(msgraphgocore_test.User)
+		item := pageItem.(mocks.User)
 		res = append(res, *item.GetId())
 
 		if *item.GetId() == "2" {
@@ -131,7 +133,7 @@ func TestIterateCanBePausedAndResumed(t *testing.T) {
 	assert.Equal(t, res, []string{"0", "1", "2"})
 
 	pageIterator.Iterate(func(pageItem interface{}) bool {
-		item := pageItem.(msgraphgocore_test.User)
+		item := pageItem.(mocks.User)
 		res2 = append(res2, *item.GetId())
 
 		return true
@@ -139,14 +141,14 @@ func TestIterateCanBePausedAndResumed(t *testing.T) {
 	assert.Equal(t, res2, []string{"2", "3", "4", "10"})
 }
 
-func buildGraphResponse() *msgraphgocore_test.UsersResponse {
-	var res = msgraphgocore_test.NewUsersResponse()
+func buildGraphResponse() *mocks.UsersResponse {
+	var res = mocks.NewUsersResponse()
 
 	nextLink := "next-page"
-	users := make([]msgraphgocore_test.User, 0)
+	users := make([]mocks.User, 0)
 
 	for i := 0; i < 5; i++ {
-		u := msgraphgocore_test.NewUser()
+		u := mocks.NewUser()
 		id := fmt.Sprint(i)
 		u.SetId(&id)
