--- msgraphgocore_test/user_response.go
@@ -1,133 +1,0 @@
-package msgraphgocore_test
-
-import (
-	i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55 "github.com/microsoft/kiota/abstractions/go/serialization"
-)
-
-// UsersResponse
-type UsersResponse struct {
-	// Stores additional data not described in the OpenAPI description found when deserializing. Can be used for serialization as well.
-	additionalData map[string]interface{}
-	//
-	nextLink *string
-	//
-	value []User
-}
-
-// NewUsersResponse instantiates a new usersResponse and sets the default values.
-func NewUsersResponse() *UsersResponse {
-	m := &UsersResponse{}
-	m.SetAdditionalData(make(map[string]interface{}))
-	return m
-}
-
-// GetAdditionalData gets the additionalData property value. Stores additional data not described in the OpenAPI description found when deserializing. Can be used for serialization as well.
-func (m *UsersResponse) GetAdditionalData() map[string]interface{} {
-	if m == nil {
-		return nil
-	} else {
-		return m.additionalData
-	}
-}
-
-// GetNextLink gets the @odata.nextLink property value.
-func (m *UsersResponse) GetNextLink() *string {
-	if m == nil {
-		return nil
-	} else {
-		return m.nextLink
-	}
-}
-
-// GetValue gets the value property value.
-func (m *UsersResponse) GetValue() []User {
-	if m == nil {
-		return nil
-	} else {
-		return m.value
-	}
-}
-
-// GetFieldDeserializers the deserialization information for the current model
-func (m *UsersResponse) GetFieldDeserializers() map[string]func(interface{}, i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode) error {
-	res := make(map[string]func(interface{}, i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode) error)
-	res["@odata.nextLink"] = func(o interface{}, n i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode) error {
-		val, err := n.GetStringValue()
-		if err != nil {
-			return err
-		}
-		if val != nil {
-			m.SetNextLink(val)
-		}
-		return nil
-	}
-	res["value"] = func(o interface{}, n i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.ParseNode) error {
-		val, err := n.GetCollectionOfObjectValues(func() i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable {
-			return NewUser()
-		})
-		if err != nil {
-			return err
-		}
-		if val != nil {
-			res := make([]User, len(val))
-			for i, v := range val {
-				res[i] = *(v.(*User))
-			}
-			m.SetValue(res)
-		}
-		return nil
-	}
-	return res
-}
-func (m *UsersResponse) IsNil() bool {
-	return m == nil
-}
-
-// Serialize serializes information the current object
-func (m *UsersResponse) Serialize(writer i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.SerializationWriter) error {
-	{
-		err := writer.WriteStringValue("@odata.nextLink", m.GetNextLink())
-		if err != nil {
-			return err
-		}
-	}
-	{
-		cast := make([]i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable, len(m.GetValue()))
-		for i, v := range m.GetValue() {
-			temp := v
-			cast[i] = i04eb5309aeaafadd28374d79c8471df9b267510b4dc2e3144c378c50f6fd7b55.Parsable(&temp)
-		}
-		err := writer.WriteCollectionOfObjectValues("value", cast)
-		if err != nil {
-			return err
-		}
-	}
-	{
-		err := writer.WriteAdditionalData(m.GetAdditionalData())
-		if err != nil {
-			return err
-		}
-	}
-	return nil
-}
-
-// SetAdditionalData sets the additionalData property value. Stores additional data not described in the OpenAPI description found when deserializing. Can be used for serialization as well.
-func (m *UsersResponse) SetAdditionalData(value map[string]interface{}) {
-	if m != nil {
-		m.additionalData = value
-	}
-}
-
-// SetNextLink sets the @odata.nextLink property value.
-func (m *UsersResponse) SetNextLink(value *string) {
-	if m != nil {
-		m.nextLink = value
-	}
-}
-
-// SetValue sets the value property value.
-func (m *UsersResponse) SetValue(value []User) {
-	if m != nil {
-		m.value = value
-	}
-}
